Remove empty miniatures.
poo#23422.

When an invocation to assert_script_run doesn't change the screen, the test result would
allocate the miniature for the screen capture, but since the screen didn't change, it
would became a empty space, causing the webinterface to overlap both areas.

The change now "allocates" the space only when the screen can be captured.