Polling the read socket in the capture loop
This change will make the capture loop more robust by polling the read socket to
detect half-open connections
When a connection is closed only on the other end if we try to read from it, an
error isn't yielded (not like how it happens with writes), creating an infinite
capture loop

There are three ways to solve this problem:
+ Send a message either to the application protocol (null message) or to the
  application protocol framing (an empty message)
  Disadvantages: Requires changes on both ends of the communication.

+ Polling the connections (Note: This is how HTTP servers work when dealing with
   persistent connections)
   Disadvantages: False positives

+ Change the keepalive packet settings
  Disadvantages: TCP/IP stacks are not required to support keepalives.

The first and last option, are a bit troublesome as we need to deal with restarts.
The first option also requires changes in both ends - Besides having to
enable/disabe the SIGPIPE capture on demand, when we have a SSH connection we
start hitting the error "bad packet length".