qemu: Increase memory dump performance
Makes three changes to the memory dump:

1) Uses the file descriptor URI with the migrate command instead of exec. This
means that QEMU is writing the data directly to a file instead of writing it
to a pipe to /bin/sh which then relays it to gzip which tries to (re)compress it
and write it back to a pipe to sh which then writes it to a file.

QEMU can compress the VM memory internally using zlib, however the resulting
image can not be read by the crash tool. So by default we compress the
resulting file using xz (or bzip2 if xz is missing). The internal QEMU
compression can still be used, but the user can probably only use it to
perform an incoming migration.

Two new variables are introduced to allow control of the compression method
used.

2) Set max bandwidth to the largest size QEMU will accept to make sure that
the transfer speed is not being limited by QEMU. It seems unlikely that this
is a problem for us, but I am not sure if QEMU can set this value
automatically.

3) Make sure the VM is frozen before beginning the migration. QEMU will try to
copy most of the RAM while the VM is still running which requires it to track
dirtied RAM and resend it. Eventually it freezes the machine to finish the
copy. However if the guest is busy then this could increase the size of the
dump and cause it to take longer. It also means QEMU has to maintain a bitmap
of the guest RAM to track modified regions.