tinycv: Add send_with_fd utility function
Avoid using Socket::MsgHdr by implemented the needed functionality in a simple
utility function. Unfortunately this can't be done in Perl very easily, so it
is implemented in C. Because of the way the build system works (including
runtime loading), it is simplest to put the code in tinycv although it is not
necessarily related to its other functionality.

Note that Socket::MsgHdr does not compile correctly when used in os-autoinst
because we are wrapping everything inside eval(s). Also there is a cost
associated with each new dependency which becomes significant in a large
project.