testapi: Ensure script_output only returns the actual ouput of the executed script
This prevents additional data on the used serial device to confuse
script_output callers which do not expect script_output to implicitly consume
and return that additional "noise" data.

The test data mocking in the unit is certainly not perfect because it relies
on the actual definition of context markers used within the implementation of
script_output but I have not found a better way.

Related progress issue: https://progress.opensuse.org/issues/30613