testapi: Refactor script_output into distribution.pm and fix race (#926)
script_output's implementation is very specific to the Operating System and
even to a particular Linux distribution. This moves it into a place where it
can be overridden by the distribution and can also use distribution specific
variables more easily.

This also fixes a race condition with the serial terminal console by waiting
for the shell prompt to appear.