Add support to override HDDMODEL per disk
HDDMODEL can be set to 'nvme' to make qemu use this disk type. However,
sometimes it is desired to just have this extra NVMe disk in addition to
a classical setup.
This commit adds a a new variable HDDMODEL_$i available for all disks.
When it is set, the specified disk will be made availabe as NVMe-disk,
and the snapshot-feature will be disabled, like with the global
HDDMODEL='nvme'.
The HDDMODEL_$i will override the global HDDMODEL, but does not replace
the parameter in order to not break anything.

Signed-off-by: Michael Moese <mmoese@suse.de>