Search for patterns in the SERIAL file and raise failures
With these changes the tester can now define if he wants to raise a soft or a hard
failure when a pattern appears in a test.

There's two ways to use this functionality:
1- On main.pm the developer after setting the distribution, can set the serial failures interface
   Example:

> $testapi::distri->set_serial_failures(soft=>['gcc version 4.8.5 (SUSE Linux)'], hard=>[qr/No iBFT detected/]);

2- On the tests the developer needs to set the serial_failures variable of the test
   object

The patterns defined in main.pm file are going to be available to all tests, the
developer can then override those patterns by setting the instance variable.

The hashref (case 2) needs to have this structure:
\%SERIAL_FAILURES = (
                 soft => [quotemeta "pattern 1", qr/pattern2(?!pattern3)/],
                 hard => [qr/pattern 3/]
);

The SERIAL_FAILURES keys needs to be soft or hard.
The "soft" key will raise a softfail warning in the test.
The "hard" key will raise a hard fail in the test.

Both keys need to have an arrayref containing strings.
The strings represent the patterns to look in the serial file.

The comparation will be done line by line (please pay attention to the $/ var)
and comparing the pattern string/regexp as-is with the line.