Fix missing text thumbnail after needle related test failures
This fixes a regression that needle related test failures did not create a
message "# Test died: ..." in the logfile as well as no text thumbnail with
information and what happened. This regression must have been introduced about
at least a year ago. I could not go further in my bisecting without needing to
come up with test changes.

It seems that the original commit 1759b5e introducing the exceptions has
caused this as an oversight.