Correct return values for match_has_tag
Happened to notice today the explanation of `match_has_tag`'s
return values is, strictly speaking, incorrect. It only returns
undef if no needle match has yet taken place when it is called.
If there is a most recent needle match but it does not have the
specified tag, it returns 0, not undef.

Signed-off-by: Adam Williamson <awilliam@redhat.com>