Avoid conflicting block device options
Only specify the final overlay in a block device chain and rely on the backing
files to be read from the qcow2 image. Otherwise we get errors from QEMU as
block devices specified as backing files try to inherit the overlay's
properties.

Only specifying the last block in the chain prevents raw format from being
used as an overlay. However this is an unlikely scenario.