Use bootindex with non-uefi systems
Even non-uefi bios's support the bootindex device parameter now, which is more
flexible than the old '-boot order=<drive>' command line option. In particular
it supports setting USB devices as the first boot drive which eliminates the
need for selecting them in the bios boot menu (which is unreliable as the
order can change).