Avoid resetting all consoles while reverting to a snapshot
All consoles need to be "disabled", but not "reset". Resetting consoles causes
the SUT to try logging into the them again, but they may already be logged in
at the point to snapshot was taken. Disabling just disconnects os-autoinst
from the console which is necessary while restarting QEMU. After restarting
QEMU we can reconnect by calling activate directly.

It is assumed that consoles which can't be disabled don't need to be
reënabled.