Treat ssh_console as serial_terimanl
Set ssh_console to is_serial_terminal == 1. And use
virtio_screen as duplex communication.

NET::SSH2 sock can be used with select, but we still need
the NET::SSH2:Channel for read/write.
Added a optional paramter to specify the fd which will be used
in select.

Redirect STDERR output to STDOUT to catch everything in read_until().