Block and ignore SIGTERM from opencv threads
opencv opens a lot of threads, that do not have a perl interpreter. But
if isotovideo sends a TERM to the backend process, it's undefined which
thread receives it and so we might run into this old perl5 bug:

https://rt.perl.org/Public/Bug/Display.html?id=120951

To avoid this, block SIGTERM from being received before loading opencv.
Afterwards unblock SIGTERM so the main thread can actually receive to
e.g. shutdown qemu