Ensure main.pm is still used for proper initialization on SCHEDULE
For os-autoinst-distri-opensuse the main.pm does the initialization of the
distribution which we need to execute most of the test modules successfully.
When defining the schedule using the test variable SCHEDULE we therefore need
to load the main.pm but ignore all "loadtest" calls in there.

Verification run: http://lord.arch/tests/1886