Use video captions to overlay wallclock time
Although durations and absolute time can be somewhat derived using the
framerate (it is captured with nominally 2 fps), this is cumbersome and
not always correct. There are occasional frame drops due to stalls or long
running screen matches, especially on e.g. aarch64.

Create a caption file which shows the same time stamps as the autoinst.log,
these can be overlaid over the actual video.

See poo#42050 for details.