Correct HDDMODEL / CDMODEL deprecation message (#1064)
The current code produces a message like this:

"Set HDDMODEL to scsi-h ..."

If it was CDMODEL, the message would say "scsi-c". In both cases
it's missing a 'd'. The old code which actually parsed this form
added a 'd' after the letter that's parsed out of the variable
name; when the parse code was turned into the deprecation message
code, that 'd' got lost.

Signed-off-by: Adam Williamson <awilliam@redhat.com>