Don't write _SECRET_* vars in vars.json
If a variable name starts with _SECRET_*, it is excluded from
writing in to the final vars.json file.
During the JOB, the variable is accessible via the normal functions
e.g. get_var().

Why: This is needed to store sensitive data (credentials) in workers.ini
and prevent them to be visible via the web-ui.