Add support to checkout git repos+refspecs
As a test developer I want to execute an openQA test on a production
server based on a pull request to a test distribution to not need a
local openQA instance.

So far I tested this with the included integration test as well as
manually by specifying a real valid remote test distribution and branch
which has shown to be successful. The next steps to test would be
including the full openQA stack with remote workers including their
caching setup.

Related progress issue: https://progress.opensuse.org/issues/44327