Add python binding/import/wrapping code for testapi
This is based on bmwiedemann's work, see for example the initial idea in
https://github.com/os-autoinst/os-autoinst/pull/364 .
The idea for "python bindings" was recently brought up as well in
See https://trello.com/c/GSXjk2wZ/35-python-bindings-for-openqa

With this test modules can written in python code importing the testapi
module from python. All perl objects not within the testapi can still be
accessed using the dynamic python module "perl".

Detailed changes:

* 08-autotest.t: Add test for python test modules
* t: Add test for python test modules using perl testapi
* Add python testapi interface

This is a resubmit of
https://github.com/os-autoinst/os-autoinst/pull/1089 updated to the
current state of the project but keeping the original idea. By now
different approaches have been followed to improve the maintainability
of complex test plans. However the concerns in
https://github.com/os-autoinst/os-autoinst/pull/1089 have not been in
addressed to change the design.

Related progress issue: https://progress.opensuse.org/issues/72868