qemu: use /dev/urandom as source for virtio RNG
qemu by default uses the host's /dev/random as the source for
virtio-rng, which means we still *can* run out of entropy if a
bunch of guests demand a lot and exhaust /dev/random on the
host. Since nothing run inside os-autoinst should be used in
any kind of security-sensitive production context, let's just
use host /dev/urandom as the entropy source for os-autoinst VMs
instead.

For a case where this caused some problems, see:
https://bugzilla.redhat.com/show_bug.cgi?id=1663318

Signed-off-by: Adam Williamson <awilliam@redhat.com>