Use rsync with checksum to copy installation mediums
rsync always checks if copied files' checksum matches with original
file, but we might have the case that when we retrigger disk creation,
there is medium with matching modification date and size, but with
different content. Decision based on checksum is more reliable.

See [poo#41414](https://progress.opensuse.org/issues/41414).