svirt: Pass credentials via %hash in run_ssh_cmd()
Move $hostname, $password args to %args hash.

This is a preparation for a next commit, where additional options
for %args will be added.

TODO: refactor selecting variables in read_credentials_from_virsh_variables()
according to whether used for connecting to VM host (VIRSH_HOSTNAME) or
intermediary which handles VNC (VIRSH_GUEST). Then it'd be possible to
use it for run_ssh_cmd() as well (and remove duplicity in load_snapshot()).