Allow multiple tap interfaces in one network
There is the need to have 2 network cards in one Network (e.g.
bonding).

The idea of this patch is to specify the number of nics using the TAPDEV
variable, but use the automated mechanism of openQA to decide the
tapdev name.

If the VLAN tag isn't specified, the last specified VLAN-Tag will be
used. This ensure, if openqa-ui allocates a VLAN tag to that job,
the tag is used for both interfaces.