Use default graphics on x86 to circumvent "cirrus" limitations (#1169)
"cirrus" was the default graphical adapter since 8 years (see e3ac9c27) or
longer, it is time to trust the qemu default option better now :)

This uses "std" on qemu >= 2.2 which we can assume to have on usual
systems nowadays. As we do not detect the qemu version we should simply
rely on the default on all versions.

This was already suggested e.g. in
https://bugzilla.suse.com/show_bug.cgi?id=1064041#c19

What about the alternatives?:

* *qxl*: Might provide better performance but is also more recent.
Still, we have problems with it in drivers, e.g. see
https://bugzilla.opensuse.org/show_bug.cgi?id=1057241 just recently
fixed and only for more recent products
* *virtio*: Demands a more recent qemu without more benefit for now

See https://progress.opensuse.org/issues/53339#note-2 for verification.

Further references:
* https://wiki.archlinux.org/index.php/QEMU#Graphic_card

Related progress issue: https://progress.opensuse.org/issues/53339