sshVirtshSUT use ssh aware serial screen
The method of using NET::SSH2->sock() to do the select and then read
on the NET::SSH2::Channel wasn't reliable.
It got stock from time to time: https://progress.opensuse.org/issues/54260

This approach is to use the polling approach of an none blocking socket.

Make the ssh socked none blocking.