Fix missing data while reading from virtio_console
During some test, we figured out that data get droped while reading
from virtio_console unix socket. This happens, if a command produce a
huge amount of output. And the data can be dropped in the middle, which
has the effect, that the command is actually stated as "ok".

E.g. `cat large_file_140kb.txt` sporadically have some output gap somewhere
in the middle of the output.

Interesting is, that the logfile, which is created from qemu, contains
all data.

It seems to be a buffer size problem. But I wasn't able to set a proper
buffer size on the unix socket.

So switching to named pipes. With named pipes we are able to set the
buffer size. By default we set the size equal to
/proc/sys/fs/pipe-max-size and it can be overwritten by setting the
VIRTIO_CONSOLE_PIPE_SZ variable.