Fix in-package test failure on git
8d38d28 introduced a now fatal check if the version can be evaluated or
not. The package tests failed silently in before because git was not
available in the build context when it even should not be available
because we want to use a fixed version from the package within the
isotovideo application.

This commit moves the version replacement from %install to %prep to
adjust the version in the isotovideo source code before installing it
hence having an effect on both the installed file as well as the file
used by the tests in %check.

This fixes the problem observed in OBS with the error message:
"Can't exec "git": No such file or directory at
/home/abuild/rpmbuild/BUILD/os-autoinst-4.5.1565258100.2ea934eb/isotovideo
line 107."