Add configuration option for number of NVMe queues
There are scenarios, where we should have control over the number of
queues a NVMe device has. The number of queues varies, and recent
versions of Qemu emulate 32 queues, for example.
In some scenarios, like the one depicted in [1], we need to set the
number of queues to a value smaller than the number of NUMA-nodes in the
system.
In order to allow to test this with reasonable ressource requirements,
instead of running a machine wit 40+ NUMA-nodes, it is more reasonable
to run a test with four nodes and two NVMe queues instead.
With these changes, it is possible to configure the NVMe queues by
setting the HDDNUMQUEUES_n parameter for the disk number n.

[1] https://progress.opensuse.org/issues/55502

Signed-off-by: Michael Moese <mmoese@suse.de>