Factor out click functionality from assert_and_click
There are two cases where the fixed combination of assert_screen with
the click functionality is problematic:

1. Multiple needles, where only some need/support clicking. This is
  typically solved with a combination of assert_screen + match_has_tag +
  assert_and_click. This is wasteful, as it checks twice for the same
  needle.

2. Repeated assert_and_click due to lost events. As it is impossible to
  know if an event was lost or just handled late, the repeated
  assert_and_click may encounter the just changed screen contents.

Splitting the click allows to do it conditionally, i.e. for (1.), just
replacing the assert_and_click by click_lastmatch, and for (2.), by using
something like:
```
assert_and_click(needle);
while(1) {
  assert_screen([needle, after]);
  last if match_has_tag(after);
  click_lastmatch;
}
```
For a sequence of clicks, the "after" can specify the coordinates for
the next click.