Allow read_json() to return multiple results
The read_json in commands.pm

    Mojo::IOLoop->singleton->reactor->io

is triggered only once if there is something to read, but the buffer
might contain more than one JSON at that time.

Before the result was that the JSON items were delayed to the next
read, which caused bugs.

I changed the incremental parsing from Cpanel::JSON::XS to our own
because it was difficult to debug.

* Add $DEBUG_JSON, so it can more easily debugged if necessary.
* Add test for myjsonrpc read_json/send_json