Revert "Allow loading needles from current working directory" to fix regression
This reverts aabfa1ab and dependant changes as the updates of
screenshots on the side of the openQA webUI are not referenced by the
correct path anymore. The resulting symptoms are error messages in the
logfiles of openQA webUI as well as missing screenshots in job details.

Related progress issue: https://progress.opensuse.org/issues/58289