Mojo::Reactor: stop/start reactor around reading
If time passes after the read_json() then subsequent messages to
the socket are not triggering the io callback again.
This can be reproduced by adding a `sleep 3` after the read_json and running
full-stack.t in openQA.

stopping and starting the reactor at the start/end of the callback
fixes it.

Issue: https://progress.opensuse.org/issues/59100