Follow VNC_TYPING_LIMIT everywhere (including send_key)
Also lower the default of it, systems are just not tested with that
fast typing speed and lead to unrealistic failures. Fast human typers
can reach 40-50 *words per minute*, which would translate to ~3 keys
per second (a word is averaged to be 5 keys in these type speed contexts).

So trying 50 keys per second is very likely to fail tests that wouldn't
be reproducible in real life. So let's lower this to ten times as fast
as human - and test scenarios that require super fast typing can
increase the VNC_TYPING_LIMIT

Co-Authored-By: Guillaume Gardet <guillaume.gardet@arm.com>