Allow unsetting 'fatal' test flag without snapshot support
This change enables the workflow described in the following ticket:
https://progress.opensuse.org/issues/58166

However, it does *not* change the default behavior. So when tests
without snapshot support fail they are still stopped by default. Only
when `fatal => 0` is set via `test_flags` the job will continue. Of
course no rollback to previous snapshots is done.

Note that this also affects the QEMU backend when snapshots are not
available (e.g. `QEMU_DISABLE_SNAPSHOTS` has been set).