svirt: Wait for logfile from script in open_serial_console_via_ssh
Because of commit cc0d5e79, we do not reconnect to a remote SSH command
with each `run_ssh_cmd()` call. This change the timing and cause a problem when
running `virsh console` in a `script` session, because the first
`type_string("\n")` command comes to early and gets discarded.

This workaround waits till `script` created the logfile, so we can
assume that STDIN and STDOUT is ready to be used.