Reduce overly long waiting time on shutdown waiting on processes
This is following the perl best practice to use the most important features
from upstream dependencies which are of course undocumented ;)
The parameter "sleeptime_during_kill" from Mojo::IOLoop::ReadWriteProcess
allows to signficantly reduce the shutdown time of an isotovideo process
from about 5s to 1s. The actual actions by Mojo::IOLoop::ReadWriteProcess
can be followed by enabling the environment variable MOJO_PROCESS_DEBUG=1.
The "commands" server seems to be very slow to shut down so it might
actually receive the KILL signal after it does not react to TERM in time.
This is not changed by this commit only that we wait less until we come to
this step.

There is also a proposal for an upstream change to reduce
the initial wait time before sending termination signals, see
https://github.com/mudler/Mojo-IOLoop-ReadWriteProcess/pull/6

Related progress issue: https://progress.opensuse.org/issues/58379