Remove @INC modifications from tests
Tests should not modify @INC themselves (and if it must, then via FindBin to
use absolute paths).
The paths should be provided via env vars or `prove -l` or `prove -I....`.
This makes it more flexible.

Also the paths were relative, so it was only possible to call the tests
from the t/ directory.

This is a part of a refactoring regarding this issue:
https://progress.opensuse.org/issues/42074

In t/00-compile-check-all.t the `use lib` is still necessary because
we are passing a relative path via PERL5LIB.