t: Enable 10-terminal.t to be run with prove
Enabling test harness, for example with prove, requires the test numbers
to be subsequent and in order.

Because this test is running test functions in child processes, the
order is random, and prove will fail.

Usually, Test::SharedFork is used to create ordered test numbers.
But it does not work for this test. Maybe there is more fork magic
going on.

Therefor, we collect the test functions in the child processes via a
shared JSON file and run them when the childs finished.

To see the difference, run this command in master and with this PR:

    (cd t && prove -I$PWD/.. -v 10-terminal.t)