Allow relative "productdir" and "needledir" relative to "casedir"
This might help with being more flexible regarding also custom git
checkouts in openQA as well as in the case of manual calls to isotovideo
not needing the user to provide productdir and/or needledir as absolute
paths when the general expectation is that they reside within casedir
anyway although this is not necessarily required. There was already support
for a product directory relative to the current working directory.

Related progress issue: https://progress.opensuse.org/issues/56789