Ensure tianocore boot menu is shown for aarch64 workaround
https://github.com/os-autoinst/os-autoinst/pull/1275 removed the boot
menu by default to save 5s of boot time when the menu is not really
required. With the variable BOOT_MENU we can always request the menu
when it is necessary. On aarch64 to boot a non-default disk the
tianocore menu is required so for convience we can propose a different
default value to show the boot menu unless requested differently.

Related progress issue: https://progress.opensuse.org/issues/61910