Improve log output of test module as well as lib function lines
Previously there was a log line called for every call of a test API function
but limiting the debug output line to the filename and line number within a
test module with both a very verbose file path that is often not helpful
internal information. For example the file path in case of a caching openQA
worker would be an internally computed one not interested to test reviewers.
Also, for test modules using any kind of internal library functions not the
file location and line location within that library module file would be
reported but instead a repeated output of the same file location and line
number of the calling test module would be shown.

This commit changes the behaviour by stripping the path prefix of "CASEDIR"
which is rewritten by the openQA caching feature to relate relative to paths
that test reviewer can know from the used test distribution. The second
change is that additional to the test module location any other referenced
file within "CASEDIR" is correctly reported with filename and file number.

So for example instead of

```
[2020-02-12T14:55:11.800 CET] [debug] /home/okurz/local/os-autoinst/os-autoinst/t/data//tests/tests/boot.pm:36 called testapi::check_screen
[2020-02-12T14:55:11.801 CET] [debug] <<< testapi::check_screen(mustmatch="core", timeout=15, no_wait=1)
…
[2020-02-12T14:55:13.441 CET] [debug] /home/okurz/local/os-autoinst/os-autoinst/t/data//tests/tests/boot.pm:37 called helper::my_helper
[2020-02-12T14:55:13.441 CET] [debug] <<< testapi::assert_screen(mustmatch="core", timeout=42)
…
[2020-02-12T14:55:14.432 CET] [debug] /home/okurz/local/os-autoinst/os-autoinst/t/data//tests/tests/boot.pm:37 called helper::my_helper
```

we now get

```
[2020-02-12T15:47:57.979 CET] [debug] tests/boot.pm:36 called testapi::check_screen
[2020-02-12T15:47:57.980 CET] [debug] <<< testapi::check_screen(mustmatch="core", no_wait=1, timeout=15)
…
[2020-02-12T15:47:59.623 CET] [debug] tests/boot.pm:37 called helper::my_helper -> lib/helper.pm:24 called testapi::assert_screen
[2020-02-12T15:47:59.623 CET] [debug] <<< testapi::assert_screen(mustmatch="core", timeout=42)
…
[2020-02-12T15:48:00.621 CET] [debug] tests/boot.pm:37 called helper::my_helper -> lib/helper.pm:25 called testapi::assert_screen
```

with no redundant repetition but additional information.