Fix git hash not being able to compute when no git is available
This fixes a regression introduced with 87856568 as we try to compute
the git hash of the needle checkout directory also within
t/01-test_needle.t as we moved the functionality to needle::init. As in
before we simply want to ignore any potential problem with the git
executable completely at this point. This also covers the executable not
being existant at all.