Make point in time when a test fails in log more explicit
An openQA test that fails shows in a thumbnail popup a text like
"# Test died: no candidate needle with tag(s) 'updates_click-install' matched"
but in the autoinst-log.txt it is not really apparent and just an
innocent
"[2020-03-31T15:51:13.134 CEST] [debug] no candidate needle with tag(s) 'updates_click-install' matched"

which is not even easy to distinguish from a check_screen that is
allowed to return no matches. This should be made more prominent.

Related progress issue: https://progress.opensuse.org/issues/65106