t: Improve scope of cleanup 'chdir' call
The call "chdir $Bin" at the end of tests is done to revert the earlier
"chdir $dir" at the beginning of tests. By doing that in an "END" block
we ensure the call to be done regardless of which way the test exists as
well have both calls together so that it should be more obvious what
corresponding calls go together.