Add optional fatal handling for qemu dbus calls for easier error analysis
MM tests fail to access other machines in the network if tap device is
not present, there is a diag message reported but the test does not fail
at this point.

For example a test can show "Failed to connect to 10.0.2.2 port 20243:
No route to host" and fails. What is hidden in autoinst-log.txt:

```
[2020-05-04T08:30:16.586 CEST] [debug] Failed to run dbus command 'set_vlan' with arguments 'tap23 13' : 'tap23' is not connected to bridge 'br1'
```

Multi-machine tests relying on a tap device should fail hard if the tap
device is not present or can not be configured. With the added feature
switch QEMU_FATAL_DBUS_CALL this can be enabled. The original behaviour
had been introduced in 82fe22e2 . The reason for continuing despite the
error is unknown to me hence for now I am just adding a feature switch
for changed behaviour which we eventually might want to use as the only
new supported behaviour.

Related progress issue: https://progress.opensuse.org/issues/66376